fix: handle offline server on core apps page
