+++ src/pages/CoreApps/CoreApps.module.css
@@ -1,0 +1,3 @@
+.appHubWarning {
+    margin-bottom: var(--spacers-dp16);
+}
